Add support for native camera and QR code scanner
Known issue: the camera app that the scanner depends on has not been properly maintained lately and uses a deprecated API. This API has been recently removed from React Native, so in order to make it work we need to go and patch the camera app in node_modules. This is not ideal, but it works for now.